Moved ActorGroupManager to Utils. Add some error messages in ClassInfoManager if ActorGroups / WorkerType mappings not found
