+++ SpatialGDK/Source/SpatialGDK/Private/Utils/ActorGroupManager.cpp
@@ -1,4 +1,4 @@
-#include "ActorGroupManager.h"
+#include "Utils/ActorGroupManager.h"
 #include "SpatialGDKSettings.h"
 
 UActorGroupManager::UActorGroupManager()
